+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Server:
 - server/irc/irc-client-cache.js - Added function triggered by SIGINT and SIGTERM that asynchronously writes the cache file, then process.exit(0). The JSON encoded cache includes timestamp, cache pointer, and message data converted to an array of type utf-8 String to file `logs/savedMessageCache.json`. 
 - server/irc/irc-client-cache.js - Added javascript to run on web server restart to asynchronously read the cache file. If the timestamp age is
 300 seconds or less, the cache pointer and cache data is restored as Array of type Buffer. Cache data is then removed from the file.
-- example-credentials.json - Added new property `enableTerminateSignals: false` used to enable auto-save and restore of message cache across web server restarts. In the case of version upgrade, if the new property is missing from the configuration, the auto-save will not be enabled, but no errors are thrown for missing property.
+- example-credentials.json - Added new property `persistIrcMessageCache: false` used to enable auto-save and restore of message cache across web server restarts. In the case of version upgrade, if the new property is missing from the configuration, the auto-save will not be enabled, but no errors are thrown for missing property.
 - docs/login-config.html updated help files.
 
 Browser: